Consolidate Django apps
Merge analyze, predefined_shapes, and watershed_model into modeling.

Connects #115
