--- src/mmw/apps/predefined_shapes/fixtures/README.md
@@ -1,63 +1,0 @@
-# Data
-
-## Congressional Districts
-
-### Loading the Data
-
-The congressional district dataset, called `cb_2013_us_cd113_500k.zip`, came from [this website](https://www.census.gov/geo/maps-data/data/cbf/cbf_cds.html#cd113) and at the time of this writing can be obtained directly from [this link](http://www2.census.gov/geo/tiger/GENZ2013/cb_2013_us_cd113_500k.zip).
-The that dataset was imported into table by unzipping the just-mentioned archive into a directory on a computer that has the `shp2pgsql` command installed, then typing
-   * `shp2pgsql -s 4269:4326 -g geom cb_2013_us_cd113_500k.shp  > cb_2013_us_cd113_500k.sql`,
-   * then transferring the file `cb_2013_us_cd113_500k.sql` to the `services` VM,
-   * then typing `sudo -u postgres psql -d mmw` to get into the database,
-   * then typing `\i cb_2013_us_cd113_500k.sql` from within that program to load the data into a table called `cb_2013_us_cd113_500k`.
-
-Once the data were loaded, they were transferred to a different table (with different column names and some columns omitted) by putting to following into the `models.py` file:
-
-```Python
-class DistrictsPrimative(models.Model):
-    class Meta:
-        managed = False
-        db_table = 'cb_2013_us_cd113_500k'
-
-    gid = models.IntegerField(primary_key=True)
-    statefp = models.CharField(max_length=2)
-    cd113fp = models.CharField(max_length=2)
-    affgeoid = models.CharField(max_length=13)
-    geoid = models.CharField(max_length=4)
-    lsad = models.CharField(max_length=2)
-    geom = models.MultiPolygonField(null=True)
-
-    def __str__(self):
-        return str(self.gid)
-
-def transmute(dp):
-    return District(state_fips=dp.statefp, district_fips=dp.cd113fp,
-                    aff_geoid=dp.affgeoid, geoid=dp.geoid,
-                    lsad=dp.lsad, polygon=dp.geom)
-
-class District(models.Model):
-    state_fips = models.CharField(max_length=2)  # state FIPS codes
-    district_fips = models.CharField(max_length=2)  # 113th congress FIPS codes
-    aff_geoid = models.CharField(max_length=13)  # american fact finder geoid
-    geoid = models.CharField(max_length=4)  # geoid
-    lsad = models.CharField(max_length=2)  # legal statistical area description
-    polygon = models.MultiPolygonField(null=True)  # district polygon
-
-    def __str__(self):
-        return str(self.district_fips)
-```
-
-then generating and running the appropriate migrations, running the Django management command `shell`, and typing
-
-```Python
-from apps.home.models import (DistrictsPrimative, District, transmute)
-for dp in DistrictsPrimative.objects.all():
-    d = transmute(dp)
-    d.save()
-```
-
-into the python REPL.
-
-### Generating A Fixture
-
-The fixture was generated by typing `./scripts/manage.sh dumpdata --format=json home.district > district.json`.
