+++ src/mmw/mmw/urls.py
@@ -12,8 +12,7 @@ import watchman.urls
 import rest_framework.urls
 
 import apps.geocode.urls
-import apps.watershed_model.urls
-import apps.analyze.urls
+import apps.modeling.urls
 import apps.home.urls
 import apps.home.views
 import apps.water_balance.urls
@@ -34,8 +33,7 @@ urlpatterns = patterns(
     url(r'^watchman/', include(watchman.urls)),
     url(r'^accounts/', include(registration.backends.default.urls)),
     url(r'^api/geocode/', include(apps.geocode.urls)),
-    url(r'^api/watershed_model/', include(apps.watershed_model.urls)),
-    url(r'^api/analyze/', include(apps.analyze.urls)),
+    url(r'^api/modeling/', include(apps.modeling.urls)),
     url(r'^simple/', include(apps.water_balance.urls)),
     url(r'^user/', include(apps.user.urls))
 )
